Changes to structure of code. Now using typeProviders

+++ Web.NetArea/Point.fs
@@ -1,5 +1,5 @@
 ﻿module Point
-type Point = float * float
+type Point = decimal * decimal
 
 
 let PointToString (point:Point) =
